requirements.txt
sqlalchemy
flask-sqlalchemy
python-dotenv
pandas
psycopg2-binary